UI improvements.
We can now navigate the capture view using W,A,S,D.  We can now use the arrow keys to move the
selected textbox (left, right, up, down). Threads in the capture view now have an identifier with thread
name on the left of the capture view. We can click and drag the on the identifier block to re-order the
threads as we please.