Stack sampling with perf_event_open and unwinding with libunwindstack
Stack sampling (and for now stack unwinding) is done on the same thread as the
collection of scheduling switches.

Until we are able to always build on the host, this contains a few hacks to
build Libunwindstack on the gamelet: we replace a few files in order to remove
Android dependencies or the usage of C++17 features.