Added GGP integration classes.
GgpClient talks to the GGP-CLI executable.
It has to be in the path of the environment.

GgpClient can retrieve a list of GGP instances,
modeled by the GgpInstance class. This is basically
a list of fields form the JSON response.

GgpInstanceItemModel is a Qt Item Model, which allows
to show a list of instances in a QTableView (or any other view).

Check out ggpinstanceitemmodel_demo for how to use it.