Dynamic instrumentation with perf_event_open
Replace BpfTrace with perf_event_open for dynamic instrumentation, still using
u(ret)probes under the hood.
Also get a stack sample at the beginning and end of a dynamically-instrumented
function.

Note that stack unwinding is currently broken when a time-based sample falls
inside a dynamically-instrumented function, because of how uretprobes work.
This is not a direct cause of this change as it also happens with BpfTrace.
We have a possible fix in mind for this that will come later.