Make unwinding of samples falling inside uprobes-instrumented functions work
Uretprobes causes unwinding of stack samples falling inside uprobes-instrumented
functions to stop when the first of such functions is encountered. We solve this
by processing u(ret)probes and stack samples in order and keeping a stack of
the samples collected at the beginning of uprobes-instrumented functions to
rebuild such broken callstacks. See LinuxUprobesUnwindingVisitor.h for more
details.