TracerThread refactor to allow processing of >100K events/second. (#198)
* TracerThread refactor to allow processing of >100K events/second.

The processing loop for perf events consisted of consuming the data from the
ring buffers and processing them in place (dwarf unwinding, callbacks to
Orbit level code, etc.).  This change decouples the "consume" and "process"
parts of the loop by offloading the processing of events to another thread.
This is done to consume events in the ring buffers as fast as possible to
minimize the number of missed events due to buffer overflows.

Regarding the new deferred events, see TracerThread's DeferEvent,
ConsumeDeferredEvents and ProcessDeferredEvents methods.

Misc:
- Reduced sleep duration in the main loop from 10ms to 1ms
- Reduced size of stack copies on sampling events from 64K to 4K
- Refactored massive switch statements into function calls

Note: this is still not 100% bulletproof but it drastically improves the
number of supported callstack events from the order of 100s/second to 100Ks/second.

Tested on instance and also locally using OrbitTest.  OrbitTest is an application
that spawns N number of threads that calls a function that calls itself recursively
R times and busy works for T microseconds.  To run, call "OrbitTest N R T" with
different values.  I used mainly "OrbitTest 10 1 100" to generate >100K events/second.
In Orbit, you can select the "TestFunc" and "TestFunc2" methods of the OrbitTest
process as a benchmark.

* Revert SAMPLE_STACK_USER_SIZE to original value.

* Address PR comments.

- Make sure we handle all deferred events when stopping capture
- Revert to original implementation of  processing fork events
- Many readability changes

PR: https://github.com/pierricgimmig/orbitprofiler/pull/198/

Co-authored-by: Pierric Gimmig <pierric@google.com>