Set IsRemote flag for remote processes
Set remote flag to true for processes received from remote server.
Previously client ignored that flag of the list of modules was
empty - which is always was since prior 02840a8df9525ff85e30e10041aeb2a8508817ba
the list did not preserve information about loaded modules.

After the change the modules list is preserved and this led to a
situation where the client is incorrectly treating remote processes as
local ones.

This change fixes this problem by setting IsRemote flag on the List
after reading the process list.

Test: build, capture, restart client, capture