Fix offset in capture view by enabling uretprobes before uprobes
If we enable uprobes first, the kernel will start writing them to
the ring buffer before we have enabled the uretprobe.  When we
consume those peerless uprobes, we'll treat them as nested functions.
Start uretprobes first because we support having uretprobes without
uprobes at the beginning of a capture.

This change was Pierric's with #224. Adapted to this PR as this is going in
first.