Added simple Kokoro presubmit configs
For the presubmit builds to work we need not only supply
a Kokoro job config, but also build config files which
reside inside of the github repo.

This PR adds these file in a very simple way. No artifacts
will be generated. This needs more work in the future,
when we want to show test results from presubmit builds.