Record return address on uprobes and patch time-based samples
When a uprobe is hit, the return address is at the top of the stack.
Have perf_event_open record it as it is about to be hijacked by the uretprobe.
It will be needed to patch time-based samples (the logic is in
UprobesReturnAddressManager), as their stacks have the return addresses
hijacked, interrupting unwinding at the innermost dynamically-instrumented
function.

This replaces what we were referring to as "callstack stitching", so that we
don't copy the entire stack on every uprobe, but only the top eight bytes. As
each uprobes event is now only 88 bytes in total instead of 64 KB, this should
prevent losing events. UPROBES_RING_BUFFER_SIZE_KB is hence also decreased.

Tested with Infiltrator by hooking the first ~20 highest exclusive times.
The number of u(ret)probes events is ~500k/second (250k uprobes,
250k uretprobes), no event is lost. Callstacks seem alright.