Discard samples falling directly inside u(ret?)probes code
This is rare but can happen when instrumenting tiny functions called very
frequently.
While this seems like throwing away information that might be useful,
when such samples are unwound successfully, the result of unwinding is wrong
(and weird, alternating one frame looking correct and the next at an unknown
location).
Note that without patching the stack with saved return addresses from uprobes
those samples are not unwound successfully in the first place, but it's possible
that with this kind of special samples stack patching should be done
differently.
Possibly not worth investigating deeper though.
See b/153026037.