Grey out all context switches from other processes
Only color context switch slices related to the same process as the target.
Note that when selecting a (grey) context switch slice associated with another
process, all and only the other context switch slices associated with the same
thread are still colored.

A PID field has been added to ContextSwitch and Timer and this breaks
compatibility of the communication with previous builds.

Bug: b/152363779