Made FindFilesystem.cmake work with gcc-8's libstdc++
gcc-8's libstdc++.so contains experimental filesystem
symbols which make the linker happy but don't match
the definitions in the header files. According to the
devs it is necessary to force linking against libstdc++fs.a

This change requires the std::filesystem test program not only
to compile but also to run properly. In the GCC-8 case it segfaults
and the detection will choose linking against libstdc++fs.a
over not linking.