Refactor debug symbol loading and revive "session" loading.
Symbol loading code is now centralized in the new "SymbolsManager".
It uses the new "TransactionManager" to send a symbol-loading
request to the service and receive its response.  Note that the
symbols found locally are no longer sent to the service.  This can
save a lot of transfer time.  The service doesn't need to know about
debug symbols since all the required information is provided by the
UI when starting a capture.

A "session" saves the selected functions and their associated module
names so that a user doesn't have to manually reselect everything on
a new capture.  When loading a session, Orbit sends a symbol-loading
request to the service and automatically selects functions for
hooking when receiving the response.