Add TransactionManager to formalize UI/service message passing.
The transaction manager is responsible for coordinating requests from
the client (UI) to the service and responses from the service back to
the client. The goal is to centralize communications between client and
service to ensure that only *one* transaction is in flight at at any
given time. The user can enqueue transactions having a guarantee that
they will be executed in order. Note that enqueueing requests is
thread-safe.