Add SymbolsManager to centralize debug symbol loading in one file.
Symbol loading code is now centralized in the new "SymbolsManager". It
uses the new "TransactionManager" to send a symbol-loading request to
the service and receive its response.  Note that the symbols found
locally are no longer sent to the service.  This can save a lot of
transfer time.  The service doesn't need to know about debug symbols
since all the required information is provided by the UI when starting a
capture.