Profile using local symbols without sending all symbols to service.
We now send a list of Function objects to instrument rather than
sending a list of addresses. The service doesn't need any more
information than what is contained in that list to dynamically
instrument functions.  We don't need to send debug symbol information
for the entire module anymore. This saves a lot of data transfers.
From >100MB on every capture in Unreal we're down to maybe a few KB.

Functions are now stored as shared_ptr instead of raw pointers in "Process"
so that we can easily generate a list of selected functions and serialize it
using cereal (cereal doesn't work with raw pointers).  This impacted quite a
bit code.