Fix crash on "Hook" from a CallStackDataView
The problem was that `CallStackDataView` is a subclass of `FunctionsDataView`
and `CallStackDataView::GetFunction` returned a "dummy" `Function` (e.g, with
`pdb_ == nullptr`) when the row selected corresponds to an unknown function.
Calling `Function::Select` on such dummy `Function` would cause segmentation
fault.
Let's return a `nullptr` instead of a dummy so that we can check for it even in
`FunctionsDataView`.
Let's have a different function to return a dummy when necessary.

Bug: b/153149392

Note that for a better fix we would probably want a `CallStackDataView` that
is not a subclass of `FunctionsDataView`, so that we can have a behavior
independent from FunctionsDataView, include some feature like context menu
action "Load Symbols" from `SamplingReportDataView`. But let's fix this first.