Get RAX on uretprobe to track integer return values.
Added a temporary --show_return_values flag to set on the Orbit UI
to enable the feature.  The value of RAX will be displayed with the
function name and duration in the time slices.  The flag will be
removed once we have a way to specify if we should show the return
value on a per function basis.