Transactions, symbol loading and sessions (#329)
* Add TransactionManager to formalize UI/service message passing.

The transaction manager is responsible for coordinating requests from
the client (UI) to the service and responses from the service back to
the client. The goal is to centralize communications between client and
service to ensure that only *one* transaction is in flight at at any
given time. The user can enqueue transactions having a guarantee that
they will be executed in order. Note that enqueueing requests is
thread-safe.

* Add SymbolsManager to centralize debug symbol loading in one file.

Symbol loading code is now centralized in the new "SymbolsManager". It
uses the new "TransactionManager" to send a symbol-loading request to
the service and receive its response.  Note that the symbols found
locally are no longer sent to the service.  This can save a lot of
transfer time.  The service doesn't need to know about debug symbols
since all the required information is provided by the UI when starting a
capture.

* Revive "sessions" to automatically select process and load presets.

A "session" saves the selected functions and their associated module
names so that a user doesn't have to manually reselect everything on a
new capture.  When loading a session, Orbit sends a symbol-loading
request to the service and automatically selects functions for hooking
when receiving the response.

Co-authored-by: Pierric Gimmig <pierric@google.com>