Added GgpClient class
GgpClient talks to the GGP-CLI executable.
It has to be in the path of the environment.

GgpClient can retrieve a list of GGP instances,
modeled by the GgpInstance class. This is basically
a list of fields form the JSON response.

This class should be tested via integration tests as soon as we
have a framework for that.