Rewrote windows helper scripts in powershell
build.bat and bootstrap-orbit{,-ggp}.bat have been replaced by
powershell scripts build.ps1 and bootstrap-orbit{,-ggp}.ps1.
The old scripts still exist for compatibility reasons. They
forward the call to the new scripts and can be removed in the future.

This change happens in anticipation of more complex logic in the
bootstrap-scripts. I'm going to require a HTTP-request which is easy
on Linux with curl but just not possible to do on Windows Batch
in a portable way.