Added OrbitSsh target, Socket and ResultType types
Socket is a wrapper for linux sockets and Winsock.
It abstracts and supports the functionality needed for ssh.
The file Socket.cpp contains common code, whilst LinuxSocket.cpp
and WindowsSocket.cpp contain platform specific code.

There are unit tests for the Socket class

ResultType is a function return type for functions that can be used
in non-blocking mode. The value kAgain indicates it the function
would block (when in blocking mode). This means it needs to be
called again with the same arguments to return successfully.