Added TunnelManager class
The TunnelManager class provides the ability to tunnel
tcp connections via ssh ssh. It uses a local socket to connect
a local tcp connection on a port with a direct tcp channel to
a remote server. It manages the lifetime of the socket and
channel. Tick needs to be called periodically to send and
receive packages.