Replace atomic counters in BlockChain by properly aligned uint32_t's.
The BlockChain data structure is not thread safe.  As such, the atomic
counters were not necessary to begin with and slowed down insertion
significantly.  In benchmarks, removing atomics make the test run in
a third of the time it takes with the atomics in place.  Also, having
those uint32_t counters properly aligned gives us atomicity anyway.