Discard messages from previous capture.
Make sure that messages from a previous capture are not processed as valid messages. This can occur if we have buffered or in-flight messages as we start a new capture. Simply discard those messages and warn. This rarely happens, but can be triggered by button mashing (starting and stopping captures rapidly).