Introduce Capture::FinalizeCapture, State, GState
This is to add support to `Capture` for the ack sent by the service when done
sending all the capture data. In particular, it allows the capture view to keep
scrolling while more data is still coming.
In the process, some logic that was moved from `Capture::StopCapture` to
`OrbitApp::OnCaptureStopped` in a previous commit is brought back to `Capture`
in `FinalizeCapture`.
Also, the use of `GTimerManager` in `Capture::StartCapture`,`StopCapture`,
`IsRecording` is removed.
Bug: http://b/157973915