Moved Frame-Pointer Validation to GRPC (#700)
Moved frame-pointer-validator service and UI from Asio to gRPC.
Also reduces state to be present for the validation on the server (module information/functions) and instead send the functions to be loaded directly (offset and size).

Dissassembling requires the bit'ness of the process (i.e. 64 or 32 bit). This information is available on the client, but we do not want to rely on this, as the service should be able to be used e.g. as a static analysis at some point later. The bit'ness is part of the ElfHeader and as we have the modules path, we can easily read it. This PR extends the ElfFile with a method to retrieve that info.