Increase the size of perf_event_open ring buffers
There are still situations where we can lose perf_event_open events. Currently
this affects context switches in particular, but in some cases u(ret)probes
events and samples are also lost. After investigating this, it seems that this
happens when `TracerThread::Run`'s thread is not scheduled for a few
milliseconds. Even if the ring buffers were empty, they can overflow in this
time interval when they are not read.
To mitigate this, let's increase the size of ring buffers. Of course, if the
aformentioned thread is not scheduled for long enough, the same problem can
still happen, but these new values make it much less likely and seem to be
big enough for the experiments I made.
Bug: http://b/158543776