Do not require cmake as dep in lzma_sdk
We have cmake in our build profiles. Not requiring cmake in every recipe
prevents conan from downloading and building a different version of
cmake for each recipe.

Additionally we not require lzma_sdk in our conanfile. It's only a
transitive dependency by libunwindstack. But by requiring it explicitly
we can set a recipe revision hash.