Revised creation of default profiles
Default profiles are now created by including the corresponding upstream
profile. There are circumstances where this could fail but that
shouldn't happen on a modern linux distribution or any windows
installation. Additionally this fixes the "default-profiles are never
updated"-problem, since the upstream profiles can be updated with
additional options and settings which are then inherited by the default
profiles.