Move to gtest 1.10.0
This updated gtest to version 1.10.0 and make some smaller related
changes:
- gtest is now a build requirement in conan (pure semantic change)
- We don't use the gtest-provided cmake config files, rather the
integration provided by conan. This makes it more robust in cross
compilation scenarios.
- We provide our own gtest_main target which provides the same main
function implementation as gtest. This change is needed since because the
previous point does not allow us to choose whether we want to link
against the gtest-provided main function on a per-test basis.