Fix ProcessList file path
the file path of a process used to be determined by reading
/proc/<pid>/cmdline which contains the command how the process was
started. This is not absolute file path, but can be relative (./app).
Now it is determined by reading the link /proc/<pid>/exe which points to
the executable.

This functionality is moved to LinuxUtils::GetExecutablePath(pid) and
Path::ExecutableName uses it.

Also added todo comments to LinuxUtils