Move grpc_helper from gRPC package into repo
This change allows us to use the conan-generated cmake config files
instead of the grpc-package supplied ones. That simplifies the cmake
config in a cross compilation environment and allowed to remove some
hacks from the top-level CMakeLists.txt-file.

Additionally some changes to the grpc_helper function works around a bug
in MSBuild 2019 which always triggered a rebuild of OrbitProtos.