Replaced hit-count function with CodeReport class to prepare display of more information in the heat-map of disassambly- and source-code view.
Prio to this change, we passed a function to the ui that computed the
ratio of samples hit that line to samples that hit the function. As we
want to display more data, like the actual numbers and the percentages
relative to the function and the overall result, this will introduce
an interface CodeReport providing functions to retrieve these numbers.
Currently DisassemblyReport derives from this class in order to provide
the heat-map for the disassembly view.