OrbitClientGgp basic structure
Basic test in main that creates the channel (works) and tries to set
CaptureClient (error because ClientListener hasnt been implemented yet).

Main files and the structure of the project has been created and the
logic related to the grpc client will be moved onto ClientGgp, which
will implement CaptureListener as well.