Updated Ubuntu version in docker containers
This commit updates the ubuntu version from disco to eoan
in the linux docker containers with Qt support.

We encountered a bug/missing feature in Qt 5.12
(https://bugreports.qt.io/browse/QTBUG-69683) which is fixed with the
sub-release included in eoan (5.12.4). Ubuntu Disco Dingo ships with Qt
5.12.2.

Windows is not affected since we use Qt from conan on Windows in a very
recent release (5.14.4).