Fixes to build_and_upload_dependencies.sh
This fixes a bug when using bintray. A remote needs to be enabled to
authenticate against it. So we first enable the remote, do the
authentication and disable it again. The authentication will be
preserved even though the remote is disabled.

Furthermore the script accidentally contained a superfluous call to
build.sh which was replaced by directly calling conan install and conan
build.