Fix displaying of function names in the iterator UI overlay
When a function was not encountered in any samples, then the (global)
map that we maintain to map function addresses to display names did not
contain that function and the displayed string was incorrect. Whenever
we add an iterator, we insert the function address and display name into
the map now.

Add nodiscard to some functions in TimeGraph