Synchronize access to unique_callstacks_ in SamplingProfiler
This map is modified from the capture thread while accessed from the
main thread which could cause undefined behavior. This commit
introduces a Mutex that guards access to unique_callstacks_ map
Note that this is not final solution it is rather temporary fix
until we move to more safe Capture Model.