Add "searching" to the top-down view
This builds on the assumption that I don't want to actually hide (i.e., filter
out) entire subtrees if the filter is not matched—I find that confusing.
Instead, I want to expand enough so that the matching nodes are visible, but
collapse everything else. See this as a search functionality rather than a
filter functionality.