Highlight nodes matching the search in TopDownWidget
This introduces `HighlightingSortFilterProxyModel`, which inherits from
`QSortFilterProxyModel` and allows to retrieve the "result" of filter and to
"set" the nodes to highlight.
I can't disagree if you say that this is a hack. But I'm not sure what a better
way would be.
I had a look at how `ShowIncludesDialog`, `OrbitTreeModel`, `OrbitTreeItem` (now
not in the codebase anymore) were doing it. That maybe looked better, but it
brought the filtering logic all the way down to `OrbitTreeItem`. Not only is
that more complicated, but also I don't think is better conceptually, as that
logic should be further away from the model. As a matter of fact, it didn't use
Qt filtering code but custom `absl::StrContains`-based logic.