Split StartCapture in two
It is needed to dp the request to start a capture in a separate thread
to be able to stop the capture afterwards.

Before requesting a capture to the server, the Capture class is prepared and checks it is possibleto do so. It is necessary to split StartCapture in two so these errors are handled in the main thread before creating the new one.

As a result, two new methods split StartCapture in two:

+ PrepareStartCapture. To be called in the main thread.
+ RequestStartCapture. To be called in a new thread.