Update non-thread-safe data on the main thread
For global maps which are currently used from the main thread
we should make sure that during capturing they are updated on
the main thread as well, to avoid undefined behavior.

This change moves updates of capture data to the main thread.
It also introduces OnCaptureStarted and OnCaptureComplete
callbacks to the listener.

ClearData in Capture no longer resets capture state since
it is called after capture is successfully started but before
first event has arrived.