Added new kokoro build job config for clang_format
This job is a presubmit check which evaluates whether all the code
changes made in your branch are properly clang-formatted.

The check will list ALL files in the repository that need a
clang-format-run but only the changes made in the current PR are taken
into account for the success of the presubmit check.