synchronize GSelectedFunctionsMap and GSelectedInCaptureFunctions
When we start a new capture GSelectedInCaptureFunctions and
GSelectedFunctionsMap should be synchronized. However if we load capture there
could be discrepancy: GSelectedFunctionsMap filled when loading a saved capture
and not cleared when we start a new capture. It could lead to the situation when
GSelectedInCaptureFunctions is empty but GSelectedFunctionsMap is not (as it
contains data from saved capture).

Clearing GSelectedFunctionsMap will help to avoid this kind of situation and
GSelectedFunctionsMap will contain only functions that are relevant for the new
capture.