OrbitClientGgp: Start capture in a separate thread
Updated main() to use ThreadPool so the capture is started in a separate thread. Included optional argument to provide the duration of the capture in seconds; which is 10 by default.

The main thread goes to sleep and requests to stop the capture when it
wakes up. It shutdowns the thread and waits for it to finish.