Move state to CaptureClient
This change moves capture state from a global
to CaptureClient and syncronizes access to it
it is now thread-safe and consistent.

This also introduces kStarting state which indicates
that start is initiated but is not yet in the correct
state for stopping.  StopCapture waits until the state
becomes something other that kStarting before either
initiating a stop or in the case Start has failed -
ignoring user request.

Bug: http://b/161968742