Add TopDownWidget::HighlightCustomFilterSortFilterProxyModel to search
This overrides `data` which uses custom matching code to return whether an item
should be highlighted and if it matches the search. For the latter, the
user-defined `kMatchesCustomFilterRole` `Qt::ItemDataRole` is employed.
The new `ExpandCollapseBasedOnRole` expands or collapses tree nodes based on the
result of `data` with such user-defined `kMatchesCustomFilterRole`.