Add "searching" to the top-down view (#914)
Commits before squashing were messy because of rewriting it a couple of times:

* Add "searching" to the top-down view

This builds on the assumption that I don't want to actually hide (i.e., filter
out) entire subtrees if the filter is not matched—I find that confusing.
Instead, I want to expand enough so that the matching nodes are visible, but
collapse everything else. See this as a search functionality rather than a
filter functionality.

* Note about invalidating the top-down view's expansion state on search

* Highlight nodes matching the search in TopDownWidget

This introduces `HighlightingSortFilterProxyModel`, which inherits from
`QSortFilterProxyModel` and allows to retrieve the "result" of filter and to
"set" the nodes to highlight.
I can't disagree if you say that this is a hack. But I'm not sure what a better
way would be.
I had a look at how `ShowIncludesDialog`, `OrbitTreeModel`, `OrbitTreeItem` (now
not in the codebase anymore) were doing it. That maybe looked better, but it
brought the filtering logic all the way down to `OrbitTreeItem`. Not only is
that more complicated, but also I don't think is better conceptually, as that
logic should be further away from the model. As a matter of fact, it didn't use
Qt filtering code but custom `absl::StrContains`-based logic.

* Store the proxy_model_ pointer in TopDownWidget

* Add TopDownWidget::HighlightCustomFilterSortFilterProxyModel to search

This overrides `data` which uses custom matching code to return whether an item
should be highlighted and if it matches the search. For the latter, the
user-defined `kMatchesCustomFilterRole` `Qt::ItemDataRole` is employed.
The new `ExpandCollapseBasedOnRole` expands or collapses tree nodes based on the
result of `data` with such user-defined `kMatchesCustomFilterRole`.