Update OrbitClientGgp to fit refactorization
Capture and CaptureClient has changed and been refactored, so the code in OrbitClientGgp has been updated to integrate these modifications.

As a result, PrepareStartCapture() and RequestStartCapture() has been
merged into one (RequestStartCapture). Also, the new thread is not
created in main.cpp, as it is handled by CaptureClient, it is mrerely
needed to create and provide a thread pool instead.