Remove Introspection and its unused dependencies
After all the changes to the service and communication, this no longer works.
Being it behind `#ifdefs`, this doesn't get maintained.
As we are not using it and as manual instrumentation is being introduced, I
would suggest removing it.
The alternative would be to make this work again, but I'm not sure it's worth
the effort of fixing it and maintaining it.