Fix crash when selecting a different process (#997)
This bug occurred when there were symbols loaded and functions hooked.
In this case the process gets updated internally and the UI tries to
render the function window before its `OnDataChanged` event arrived,
as `UpdateModules` fires the refresh callbacks at the very end initially
spawned from another thread.

This fix pospones the setting of the global variable GTargetProcess
until all data is ready to propagete the UI update (in particular the
modules list). So if the UI redraws before the UpdateModules function
propagates the data update, there will be no inconsistent state and
thus no crash. As UpdateModules now also sets the process, the method
is also renamed. Also OnProcessSelected gets removed as it just forwards
to the new update method now.

This also slightly refactors the Capture::SetTargetProcess by making
it non-ref parameter (as it is a sink) and move the process there and
further rename the parameter.

Bug: http://b/163786513
Test: On Linux: Select a process, load some modules and hook functions.
  Then select another process. The previous crash is gone now.