Merge CI build instructions into a single script
This commit unifies the different build configurations and build scripts
used in Kokoro into a single set of scripts - one script for building
and one script for symbol uploading.

Both Windows and Linux builds are now handled by a bash scripts. Since
the Kokoro Windows environment is based on Cygwin, bash is always
available and it allows us to only maintain a single script with some
`if win32`-branches here and there.

This new script is based on the docker workflow meaning also on Windows,
building in a Docker container is required.

This commit does not yet switch the Kokoro-jobs over to the new
unified-build-scripts. This will happen in subsequent commits.