Enforce consistent formatting across different versions of clang-format.
Each version of clang-format has its own definition of Google style.
Recent fails in the presubmit check showed that there are differences
in the latest version and the one from the CI (which is clang-format-9).

This PR dumps the Google style definition from clang-format-9 into
the .clang-format file instead of relying on what the version below
has defined as such.