Refactor Capture to move GSelectedFunctionsMap to DataManager (#1026)
The GSelectedFunctionsMap global was previously used from various
places assuming different semantics, which already lead to bugs.
In order to get rid of the globals in Capture and make accesses
to the data stored there more explicit, this change moves the map
to OrbitGl/DataManager and passes it explicitly as input to OrbitCore.

This map represents the current state of the running application,
i.e. wich functions are currently hooked and is not related to the
functions hooked in the previous/loaded capture.
Now this map also stores a copy of the FunctionInfo content, to
avoid nullptrs. Making all maps to store raw pointers is unfeasable,
as there is no dedicated owner of these functions, when loading a
capture (otherwise it would be the process's module).
All accessor methods (in particual to select/unselect/check functions)
are moved to App.

Some usages were replaced with the intended selected functions from the
previous capture (i.e. in CaptureData).

Added [[nodiscard]] and const on various places.

Bug: http://b/163020637
Test: Load Modules, Hook/Unhook functions, do captures, load/store captures,
      load/store presets, repeat captures while/after hooking/unhooking function,
      filter functions.