Delete artifacts before exiting CI build script
This change makes sure that all unnecessary build artifacts are deleted
before we finish the CI build. Kokoro will copy the whole directory
before it decides which files need to go to GCS. By deleting a lot of
files upfront we will save some time, especially in the presubmit case
where only the test results are of interest and artifactory are
discarded.