Add ProcessManager instantiation to ClientGgp
ProcessManager is required to be able to handle and retrieve information
about the modules and the selected process. The refresh_timeout is
provided as an option to the client and set to default to 1000ms.

ProcessManager needs to be shutdown so it has also been created a new
method to do so: ClientGgp::ShutdownClient.