Prepare CaptureSerializer for fuzzing
The fuzzer needs access to the `ReadMessage` and `WriteMessage` function
which used to be anonymous functions living in the cpp-file. I made them
static member function of `CaptureSerializer` such that they can be
called from a different compilation unit.

Furthermore I added a `static` to another free function which was only
defined in the cpp-file and never declared in a header file.