Adjust Linux conan profiles for fuzzing
This commit adjusting the existing `libfuzzer_relwithdebinfo` profile to
the new fuzzing changes and adds a new profile `libfuzzer_base`. This
new profile does not define a compiler and is only setting some compile
flag variables. It is meant to be included by a full-profile.

This base profile is used in the oss-fuzz project where the compiler is
determined based on what's installed in the compiling docker container.

When writing this commit message it was clang-12 from clang-master.