Retrieve list of processes to set target info
To later being able to find a module, it is needed to have more information about the process than merely its pid so it is included a call to ProcessManagerImpl::GetProcessList.

The list of processes is automatically updated by a worker function in the process manager. However, for the client ggp is not required to have a full updated list of processess all the time so it is enough getting the list once to get the information about the process, which is not going to change during execution. Therefore, it is needed to wait for the list of processes to be updated at least once before continuing with the program. Once the list is populated, the callback listener function is removed and the program is able to continue.

The target process is searched in the list of processess retrieved and
its information is populated into the target_process_. An available target process
is essential for the capture to work as expected so the program ends if
not found.