Load symbols for the process binary module
The symbols corresponding to the process binary module are needed to be
loaded. The loading of the module and its symbols has been moved into a
new method called LoadModuleAndSymbols. This information is needed
during capture so it has been included to its initialisation.

After calling this method, the target_process_ has the information about
the module corresponding to its binary and its symbols.