Tutorial basics and resources
Added a framework to create interactive tutorials (tutorialoverlay.h).
Documentation on how to create new tutorials can be found there,
and in the accompanying UI file (tutorialoverlay.ui).

Also added a bunch of resources used by the tutorial system, as well
as some yet unused but potentially usefull.

Note that the design is still preliminary, we probably want to change
that in the future.