Add option to bypass the release check
We have a check which forbids to finish a Kokoro-triggered release build
when the current branch is not on a tag.

For testing purposes or in exceptional situations it makes sense to
overwrite that behaviour.

This commits adds an environment variable which disabled the check. This
environment variable can be set when the job is triggered from Fusion.