Split ProcessManager into ProcessManager and ProcessClient
ProcessManager contains the implementation of the client methods related
to the process service but also creates a worker in the background in
charge of updating the list of processes until the manager is shutdown.

Therefore, ProcessManager can be split in two so the methods related to
the client are located in their own class: ProcessClient. ProcessManager
interface does not change to the Orbit App and makes the calls to
ProcessClient instead of handling the client calls itself. This refactor
also allows using ProcessClient in other clients without having to
create a worker if it is not needed.

ProcessClient will be used in the ggp client.