Install specific version of MSVC2019 in Docker
This commit ensures that always the same version of MSVC2019 is
installed when building the docker containers instead of the latest
version. Due to regressions, defect reports or simply new features the
build can break and also did in the past. This can be avoided by
sticking to a specific version and updating when needed.

Latest example:
https://developercommunity.visualstudio.com/content/problem/1144371/failed-to-compile-llvm-9011001-in-vs-2019-1670.html