Use shared_ptr to pass Pickable
Using weak_ptr implies additional cost of having to check
if the value expired on every call in every function. This
is not necessary if we pass shared_ptr instead. In that case
the only place we have to check is when getting Pickable
from the container. After that we can lock it and pass
back to the caller.

Also changed memcpy to absl::bit_cast in ColorFromPickingID

Fixed a bug in GetOrCreatePickableId where it was not using
existing pickable_id and instead always returing id based
on current pickable_id_counter_. Added a test that catches it.

Test: run OrbitGlTests