Use ProcessClient instead of ProcessManager in ClientGgp
ProcessManager has been refactored so the methods related to the process
service are not in ProcessClient. As ClientGgp does not need the worker
the ProcessManager creates by default, it uses ProcessClient calls
directly instead.

As a result, all references to ProcessManager and the worker are removed
as well as the optional parameter process_refresh_timeout which is not
needed anymore, as it was used by ProcessManager.