Add new conan options
This commit adds two new conan options: `with_orbitgl` and `run_tests`.
Both are needed for the fuzzing workflow.

`with_orbitgl` allows to compile the OrbitGl CMake module even though
`with_gui` is False. Since OrbitGl contains fuzz-tests while OrbitQt
does not this allows to compile all fuzz-tests on a machine without Qt
installed.

`run_tests` allows to explicitly disable the test-runs in a
conan-triggered build. The normal tests might fail when the code-base
is compiled with sanitizers for fuzzing. So for the fuzzing workflow we
want to be able to not run the tests.